eliminate recursive import
